PDN: fixes #<0003161> "ShapeBinder not at expected place"
Instead of using the placement of the Part::Features the ShapeBinder
is based on, use the shapes location. The problem was, that when using
the placement of the feature, the ShapeBinder had a wrong translation
and rotation.

Discussed here "https://forum.freecadweb.org/viewtopic.php?f=3&t=23893&sid=3f16d093fcc7cc76188387dc84f46a0e&start=10"